added additional menu type
added another menu option for a menu that has clickable parent nodes for
dropdown menus.  For example, the default menu doesn't allow you to
click and go to the parent page of the dropdown item, the added menu
type allows this.

Also added info in README file on how to change menu types and added
info for changing themes.